- added lazy parameter input for basecell with x1,x2,y1,... - reverted recent changes in adjust_surface_points - started meshing boundary circles of interpretation - fixed rounding bug in optimization_tools

+++ share/python/optimization_tools.py
@@ -1079,7 +1079,7 @@ def get_image(data, scale = None):
 
   img = Image.fromarray(ret)
   if scale:
-    img = img.resize((scale, y/x * scale), Image.NEAREST) # higher quality with ANTIALIAS
+    img = img.resize((scale, int(y/x * scale)), Image.NEAREST) # higher quality with ANTIALIAS
   return img
